Extract base class out of form representer
- moving error handling and basic structure of a form into a generic base class
- leaving work package specific stuff in WP form representer
